// src/payload-types.ts
export type IRowRatio = ('50/50' | '33/66' | '66/33') | null;
// ...
export interface Media {
  id: number;
  alt: string;
  type: 'watercolour' | 'oil' | 'ceramics';
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface ISectionProps {
  variant?: ('full' | 'centered') | null;
  bg?: ('base' | 'green' | 'white' | 'greenMute') | null;
  blured?: boolean | null;
  zIndex?: number | null;
  topOffset?: number | null;
  bottomOffset?: number | null;
}